+++ internal/terminal/types.go
@@ -1,6 +1,4 @@
-// Package terminal owns long-lived PTY and SSH terminal sessions. It has no
-// transport concerns: HTTP/WebSocket adapters attach clients and forward the
-// OutputFrame stream exposed here.
+// Package terminal owns long-lived PTY and SSH terminal sessions.
 package terminal
 
 import (
@@ -51,15 +49,11 @@ var (
 	ErrAttachmentClosed = errors.New("terminal: attachment is closed")
 	ErrUnavailable      = errors.New("terminal: backend is unavailable")
 
-	// ErrBackendMissing reports that the named tmux/screen session this
-	// runtime attaches to is gone. It is never reconnectable and never
-	// re-runs the session command.
+	// ErrBackendMissing reports that the tmux/screen session is gone.
 	ErrBackendMissing = errors.New("terminal: backend session no longer exists")
 )
 
-// Credential deliberately exposes no secret serialization contract. Session
-// repositories persist a HostID and resolve the HostSpec (and credential) when
-// restoring, so private key material never needs to enter a session record.
+// Credential exposes no serialization contract, so secrets stay out of records.
 type Credential interface {
 	isCredential()
 }
@@ -101,10 +95,8 @@ type SessionSpec struct {
 	Host        *HostSpec // nil means a PTY on the wmux host.
 	Persistence Persistence
 
-	// Generation is the execution number owned by the application's session
-	// row. Runtime state callbacks carry it back so a callback from a stopped
-	// execution cannot overwrite the state of a newer one.
-	Generation uint64
+	// Generation is the execution number owned by the application's session row.
+	Generation int
 
 	Shell string
 	Args  []string
@@ -114,9 +106,7 @@ type SessionSpec struct {
 	Rows  uint16
 }
 
-// SessionRecord is safe to persist: it references a remote host but contains
-// no Credential. Active indicates that wmux should reconnect to a persistent
-// backend after a service restart.
+// SessionRecord is safe to persist: it references a host but holds no Credential.
 type SessionRecord struct {
 	ID                  string
 	Name                string
@@ -130,12 +120,10 @@ type SessionRecord struct {
 	Cols                uint16
 	Rows                uint16
 	Active              bool
-	Generation          uint64
+	Generation          int
 }
 
-// Repository is implemented by the application's storage layer. LoadHost is
-// intentionally separate from ListSessions so credential handling stays in the
-// host repository instead of being serialized into SessionRecord.
+// Repository is implemented by the application's storage layer.
 type Repository interface {
 	ListSessions(ctx context.Context) ([]SessionRecord, error)
 	LoadHost(ctx context.Context, hostID string) (HostSpec, error)
@@ -143,7 +131,7 @@ type Repository interface {
 
 type SessionStatus struct {
 	ID           string
-	Generation   uint64
+	Generation   int
 	State        SessionState
 	Persistence  Persistence
 	WriterID     string
@@ -152,9 +140,7 @@ type SessionStatus struct {
 	LastSequence uint64
 }
 
-// Callbacks run after runtime locks are released. Implementations should return
-// promptly. OnSessionState is the only persistence point for runtime state:
-// the runtime itself never writes Repository.
+// Callbacks run after runtime locks are released and should return promptly.
 type Callbacks interface {
 	OnSessionState(status SessionStatus)
 	OnWriterChanged(sessionID, clientID string)
